+++ test/dummy/config/routes.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  mount Livechat::Engine => "/livechat"
+  mount_livechat at: "/livechat"
   get "sample", to: "sample#show"
 end